Bifurcated string and object extentions methods into different files
