+++ lib/src/object_extension.dart
@@ -1,0 +1,5 @@
+/// Object Extentions
+extension ObjectExtenstions on Object? {
+  bool get isNull => this == null;
+  bool get isNotNull => this != null;
+}
